+++ package/Tests/Runtime/DummyLoader.cs
@@ -1,19 +1,18 @@
-using System;
-using UnityEditor;
 using UnityEngine;
 using UnityEngine.XR.Management;
 
 namespace ManagementTests.Runtime {
 
-    public class DummyLoader : XRLoader {
-        private bool m_ShouldFail = false;
+    internal class DummyLoader : XRLoader {
+        public bool m_ShouldFail = false;
+        public int m_Id;
 
-        public DummyLoader (bool shouldFail = false) {
-            m_ShouldFail = shouldFail;
+        public DummyLoader ()
+        {
         }
 
         public override bool Initialize () {
-            return m_ShouldFail;
+            return !m_ShouldFail;
         }
 
         public override T GetLoadedSubsystem<T>()
@@ -21,6 +20,29 @@ namespace ManagementTests.Runtime {
             return default(T);
         }
 
+        protected bool Equals(DummyLoader other)
+        {
+            return base.Equals(other) && m_ShouldFail == other.m_ShouldFail && m_Id == other.m_Id;
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((DummyLoader) obj);
+        }
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                int hashCode = base.GetHashCode();
+                hashCode = (hashCode * 397) ^ m_ShouldFail.GetHashCode();
+                hashCode = (hashCode * 397) ^ m_Id;
+                return hashCode;
+            }
+        }
     }
 
 }