+++ package/Tests/Runtime/RuntimeTests.cs
@@ -1,18 +1,51 @@
 ﻿using System.Collections;
+using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
 using UnityEngine.XR.Management;
 
 using UnityEditor;
+using UnityEngine.SceneManagement;
 
 namespace ManagementTests.Runtime {
-    class RuntimeTests {
+
+
+    [TestFixture(0, -1)] // No loaders, should never have any results
+    [TestFixture(1, -1)] // 1 loader, fails so no active loaders
+    [TestFixture(1, 0)] // All others, make sure the active loader is expected loader.
+    [TestFixture(2, 0)]
+    [TestFixture(2, 1)]
+    [TestFixture(3, 2)]
+    class ManualLifetimeTests {
         GameObject gameManager = null;
+        private List<XRLoader> loaders = new List<XRLoader>();
+        private int _loaderCount;
+        private int _loaderIndexToWin;
+
+        public ManualLifetimeTests(int loaderCount, int loaderIndexToWin)
+        {
+            _loaderCount = loaderCount;
+            _loaderIndexToWin = loaderIndexToWin;
+        }
+
+
         [SetUp]
         public void SetupXRManagerTest () {
             gameManager = new GameObject ();
             XRManager manager = gameManager.AddComponent<XRManager> () as XRManager;
+            manager.ManageActiveLoaderLifetime = false;
+
+            loaders = new List<XRLoader>();
+
+            for (int i = 0; i < _loaderCount; i++)
+            {
+                DummyLoader dl = ScriptableObject.CreateInstance(typeof(DummyLoader)) as DummyLoader;
+                dl.m_Id = i;
+                dl.m_ShouldFail = (i != _loaderIndexToWin);
+                loaders.Add(dl);
+                manager.Loaders.Add(dl);
+            }
         }
 
         [TearDown]
@@ -21,52 +54,34 @@ namespace ManagementTests.Runtime {
             gameManager = null;
         }
 
-        [UnityTest]
-        public IEnumerator CheckFirstLoaderWins () {
 
+        [UnityTest]
+        public IEnumerator CheckActivatedLoader ()
+        {
             Assert.IsNotNull (gameManager);
 
             XRManager manager = gameManager.GetComponent<XRManager> () as XRManager;
             Assert.IsNotNull (manager);
 
-            DummyLoader loader1 = new DummyLoader ();
-            DummyLoader loader2 = new DummyLoader ();
-            manager.AddLoader (loader1);
-            manager.AddLoader (loader2);
+            yield return manager.InitializeLoader();
 
-            yield return new EnterPlayMode ();
+            if (_loaderIndexToWin < 0 || _loaderIndexToWin >= loaders.Count)
+            {
+                Assert.IsNull (XRManager.ActiveLoader);
+            }
+            else
+            {
+                Assert.IsNotNull (XRManager.ActiveLoader);
+                Assert.AreEqual(loaders[_loaderIndexToWin], XRManager.ActiveLoader);
+            }
 
-            Assert.IsNotNull (XRManager.ActiveLoader && XRManager.ActiveLoader == loader1);
-
-            yield return new ExitPlayMode ();
-
-            Assert.IsNull (XRManager.ActiveLoader);
-
-            yield return null;
-        }
-
-        [UnityTest]
-        public IEnumerator CheckSecondLoaderWins () {
-
-            Assert.IsNotNull (gameManager);
-
-            XRManager manager = gameManager.GetComponent<XRManager> () as XRManager;
-            Assert.IsNotNull (manager);
-
-            DummyLoader loader1 = new DummyLoader (true);
-            DummyLoader loader2 = new DummyLoader ();
-            manager.AddLoader (loader1);
-            manager.AddLoader (loader2);
-
-            yield return new EnterPlayMode ();
-
-            Assert.IsNotNull (XRManager.ActiveLoader && XRManager.ActiveLoader == loader2);
-
-            yield return new ExitPlayMode ();
+            manager.DeinitializeLoader();
 
             Assert.IsNull (XRManager.ActiveLoader);
 
-            yield return null;
+            manager.Loaders.Clear();
         }
+
     }
+
 }